Opaque: allow immediate run of finalize handler
Introduces global flag Opaque.ForceFinalizeInMainThread. When set to
false, objects inheriting Glib.Opaque will not schedule to conclude
their finalization from withing GLib main loop in the main thread, but
will do it immediately. This can be useful in programs that don't deal
with GTK widgets, which are expected to be freed from application's UI
thread.

For backward compatibility ForceFinalizeInMainThread defaults to true.