Fix ownership of returned Opaque with GetOpaque
GetOpaque is returning Opaque objects with owner set to the
opposite value as the one requested.

This behaviour can be seen when overridingon virtual methods like BaseSink.OnQuery
that have has parameters Gst.MiniObject's and where GetOpaque is called
with owned=false.
The new Query object is owned and with an extra Ref, making it
IsWritable=false, while it should be !owned and with refcount of 1
make it IsWritable=false