Fix segfault in GException finalizer
A GError can be cleared at any point so we need to keep
a copy to keep alive while the managed GException is alive.

An example is a segfault that can be reproduced when Gst.Discoverer
post an error converted to a GException and we call the Gst.Discover
Dispose function. There is a race between the GstDiscoverer disposal
that is internally clearing the GError in the discoverer_cleanup
function and the GException's finalizer:

=================================================================
	Managed Stacktrace:
=================================================================
	  at <unknown> <0xffffffff>
	  at GLib.GException:g_clear_error <0x000a0>
	  at GLib.GException:Finalize <0x00052>
	  at System.Object:runtime_invoke_virtual_void__this__ <0x000ab>
=================================================================
The application was terminated by a signal: SIGABRT