Fix bindings using string arrays parameters
This simplifies the generated code using the GLib.Marshaller
helpers with a new IGeneratable for array strings.
It fixes a segfault that was using the mearshaller string free helper
instead of the array one.
It fixes a leak of `in` string arrays.

Only treat count paramaters that are annotated as such

This prevent all parameters starting with n_ to be treated
like array count parameters and only those that are annotated
as such

Fix method with single array length param for several arrays