Apply style format and check it in the CI
It's currently very hard to contribute to the project since there is no
consisten style. Most editors are nowadays configured to format on save,
ending with plenty of changes and having to skip to unwanted ones for
each commit.

The style format has been adjusted to impact as little as possible but
there wasn't a consistent style. Most of the changes are space-related
and the rest are line breaks on open brances.

A way to see the impact filtering most of the noise is with: git diff -w